requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.0
pep8==1.7.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
toml==0.10.2